--- src/main/java/com/KooKPaP/server/domain/member/dto/request/MemberReq.java
@@ -1,4 +1,0 @@
-package com.KooKPaP.server.domain.member.dto.request;
-
-public class MemberReq {
-}
