--- src/main/java/com/KooKPaP/server/domain/member/dto/response/MemberRes.java
@@ -1,4 +1,0 @@
-package com.KooKPaP.server.domain.member.dto.response;
-
-public class MemberRes {
-}
